daemon-cli: move status verb and shared output helpers into crate
