+++ client/doublezero/src/command/mod.rs
@@ -4,5 +4,4 @@ pub mod helpers;
 pub mod latency;
 pub mod multicast;
 pub mod routes;
-pub mod status;
 pub mod util;
